Fix visibility region sampling

+++ src/ompl/multilevel/datastructures/pathrestriction/src/FindSectionVariableNeighborhood.cpp
@@ -117,7 +117,7 @@ bool FindSectionVariableNeighborhood::variableNeighborhoodPatternSearch(
        ->getGraphSampler()
        ->setPathBiasStartSegment(head->getLocationOnBasePath());
 
-    if(depth >= magic::PATH_SECTION_MAX_DEPTH)
+    if(depth >= (int)magic::PATH_SECTION_MAX_DEPTH)
     {
         return false;
     }
